requirements.txt
streamlit
python-dotenv
pdf2image
openai